Move related-media-store to 'store'
