+++ src/store/index.js
@@ -14,7 +14,6 @@ import UserStore from '~/store-modules/user-store'
 import UsageDataStore from '~/store-modules/usage-data-store'
 import FilterStore from '~/store-modules/filter-store'
 import ReportContentStore from '~/store-modules/report-content-store'
-import RelatedMediaStore from '~/store-modules/related-media-store'
 import NotificationStore from '~/store-modules/notification-store'
 import NavStore from '~/store-modules/nav-store'
 import ActiveMediaStore from '~/store-modules/active-media-store'
@@ -35,7 +34,6 @@ export const actions = Object.assign(
   SocialMediaStore.actions(GoogleAnalytics),
   ABTestStore.actions,
   ReportContentStore.actions(ReportService),
-  RelatedMediaStore.actions(AudioService, ImageService),
   NotificationStore.actions,
   {
     async nuxtServerInit({ dispatch }) {
@@ -59,7 +57,6 @@ export const state = () =>
     ABTestStore.state,
     UserStore.state,
     ReportContentStore.state,
-    RelatedMediaStore.state,
     NotificationStore.state,
     NavStore.state,
     ActiveMediaStore.state
@@ -74,7 +71,6 @@ export const mutations = Object.assign(
   BugReportStore.mutations,
   ABTestStore.mutations,
   ReportContentStore.mutations,
-  RelatedMediaStore.mutations,
   NotificationStore.mutations,
   NavStore.mutations,
   ActiveMediaStore.mutations
