+++ test/unit/specs/store/related-media-store.spec.js
@@ -1,4 +1,4 @@
-import store from '~/store-modules/related-media-store'
+import store, { actionsCreator } from '~/store/related'
 import {
   FETCH_END_MEDIA,
   FETCH_START_MEDIA,
@@ -10,9 +10,9 @@ import { IMAGE } from '~/constants/media'
 describe('Related Images Store', () => {
   describe('state', () => {
     it('exports default state', () => {
-      const state = store.state
-      expect(state.related.audios).toHaveLength(0)
-      expect(state.related.images).toHaveLength(0)
+      const state = store.state()
+      expect(state.audios).toHaveLength(0)
+      expect(state.images).toHaveLength(0)
     })
   })
 
@@ -21,7 +21,7 @@ describe('Related Images Store', () => {
     const mutations = store.mutations
 
     beforeEach(() => {
-      state = { related: { images: [], audios: [] } }
+      state = { images: [], audios: [] }
     })
 
     it('SET_RELATED_MEDIA updates state', () => {
@@ -30,21 +30,17 @@ describe('Related Images Store', () => {
         relatedMedia: ['foo'],
       }
       mutations[SET_RELATED_MEDIA](state, params)
-      expect(state.relatedImages).toBe(params.relatedImages)
+      expect(state.images).toBe(params.relatedMedia)
     })
   })
 
   describe('actions', () => {
     const searchData = { results: ['foo'], result_count: 1 }
-    let audioServiceMock = null
     let imageServiceMock = null
     let commit = null
     let dispatch = null
 
     beforeEach(() => {
-      audioServiceMock = {
-        getRelatedMedia: jest.fn(() => Promise.resolve({ data: searchData })),
-      }
       imageServiceMock = {
         getRelatedMedia: jest.fn(() => Promise.resolve({ data: searchData })),
       }
@@ -54,9 +50,10 @@ describe('Related Images Store', () => {
 
     it('FETCH_RELATED_MEDIA on success', (done) => {
       const params = { id: 'foo', mediaType: IMAGE }
-      const action = store.actions(audioServiceMock, imageServiceMock)[
-        FETCH_RELATED_MEDIA
-      ]
+      const actions = actionsCreator({
+        [IMAGE]: imageServiceMock,
+      })
+      const action = actions[FETCH_RELATED_MEDIA]
       action({ commit, dispatch }, params).then(() => {
         expect(commit).toBeCalledWith(FETCH_START_MEDIA, { mediaType: IMAGE })
         expect(commit).toBeCalledWith(FETCH_END_MEDIA, { mediaType: IMAGE })
@@ -72,16 +69,14 @@ describe('Related Images Store', () => {
     })
 
     it('FETCH_RELATED_IMAGES on error', (done) => {
-      const audioServiceMock = {
-        getRelatedMedia: jest.fn(() => Promise.reject('error')),
-      }
       const imageServiceMock = {
         getRelatedMedia: jest.fn(() => Promise.reject('error')),
       }
       const params = { id: 'foo', mediaType: IMAGE }
-      const action = store.actions(audioServiceMock, imageServiceMock)[
-        FETCH_RELATED_MEDIA
-      ]
+      const actions = actionsCreator({
+        [IMAGE]: imageServiceMock,
+      })
+      const action = actions[FETCH_RELATED_MEDIA]
       action({ commit, dispatch }, params).catch((error) => {
         expect(commit).toBeCalledWith(FETCH_START_MEDIA, { mediaType: IMAGE })
         expect(dispatch).toBeCalledWith('HANDLE_IMAGE_ERROR', error)
